Move #[MapRequestPayload]/#[MapQueryString] test cases to own fixture
The payload mapping feature grew enough test classes to warrant a
dedicated fixture file, gated on symfony/http-kernel >= 6.3 where these
attributes were introduced (the shared symfony.php fixture has no
version guard).
